requirements.txt
# Telegram Bot
aiogram==3.15.0
python-dotenv==1.0.1
redis==5.2.1
aioredis==2.0.1

# FastAPI
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Google Sheets
gspread==6.1.4
google-auth==2.37.0
oauth2client==4.1.3

# HTTP Client
aiohttp==3.10.11
httpx==0.28.1

# Caching
cachetools==5.5.0
